Update to address clang-tidy comment

+++ offline/packages/trackreco/PHGhostRejection.cc
@@ -130,7 +130,7 @@ void PHGhostRejection::find_ghosts(std::vector<float>& trackChi2)
       auto& track2 = seeds[trid2];
       auto delta_phi = fabs(track1phi - track2.get_phi());
       if (delta_phi > 2 * M_PI) {
-        delta_phi = fabs(delta_phi - 2 * M_PI);
+        delta_phi = fabs(static_cast<double>(delta_phi - 2 * M_PI)); // address clang-tidy float->double promotion warning
       }
       if (delta_phi < _phi_cut &&
           std::fabs(track1eta - track2.get_eta()) < _eta_cut &&
